pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "calimero-client-py"
version = "0.1.2"
description = "Python client SDK for Calimero Network"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Calimero Network", email = "support@calimero.network"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "requests>=2.31.0",
    "websockets>=12.0",
    "base58>=2.1.1",
    "pydantic>=2.5.0",
    "aiohttp>=3.9.0",
    "toml>=0.10.2",
    "pynacl>=1.5.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]

[project.urls]
Homepage = "https://github.com/calimero-network/calimero-client-py"
Documentation = "https://docs.calimero.network/python-sdk"
Repository = "https://github.com/calimero-network/calimero-client-py"
Issues = "https://github.com/calimero-network/calimero-client-py/issues" 